ABI changed: hash_strmap & gold_hash_map: change load_factor from double to float
This reduce sizeof hash_strmap & gold_hash_map by 8 on 64 bit platform

+++ src/terark/hash_strmap.hpp
@@ -169,7 +169,7 @@ protected:
     FreeList  hugelist;
 	static const ptrdiff_t freelist_disabled = -1;
 
-	double   load_factor;
+	float   load_factor;
 	enum sort_type {
 		en_unsorted,
 		en_sort_by_key,
